+++ src/routes/index.js
@@ -33,5 +33,6 @@ router.use('/bracket-types', require('../modules/bracket_type/bracket_type.route
 router.use('/patients', require('../modules/patient/patient.routes'));
 router.use('/patient-alerts', require('../modules/patient_alert/patient_alert.routes'));
 router.use('/patient-hobbies', require('../modules/patient_hobby/patient_hobby.routes'));
+router.use('/patient-notes', require('../modules/patient_note/patient_note.routes'));
 
 module.exports = router;